updated result coroutine; added variables to implement combo counter
The coroutine now pulses when the player inputs the word.

I decided I wanted to add a combo mechanic when the player types consecutive perfect words. I also want to add a scoring system, though I'm not 100% sold on that idea if the scores can't be shown off in a high score table.